requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tqdm>=4.64.0
optuna>=3.2.0
mlflow>=2.5.0
kaggle>=1.5.0
